feat: add workflow to cleanup caches by a branch
This commit introduces a new workflow that cleans up caches associated with a specific branch when a pull request is closed
.

The workflow uses the `gh-actions-cache` extension to list and delete cache keys for the merged pull request branch.

This change also removes the application of the `popos.git.hooks` plugin in the `AndroidApplicationConventionPlugin`.