CI: Trigger internal and beta releases from main branch
This commit updates the CI workflow to trigger internal and beta releases from the main branch instead of using tags.

The changelog is now generated and stored in `changelogGithub.md` and `changelogBeta.md` files.

The `appDistributionUploadFullRelease` task is executed with the `--no-configuration-cache` option.

The `reckon` extension is configured to always use `PATCH` scope
.